updated scenario module
