+++ .gitignore
@@ -1,4 +1,6 @@
 venv/
 assets/voice
+voice
 ./output.mp4
-temp
+temp/*
+!temp/.gitkeep
